Shiny app for egcm
This is project we are just considering shares from just 3 sectors namely Financial Sector, Auto industry and Pharmaceutical. All the shares used are listed in NSE.
With the help of Shiny, we created an dashboard which allows you to select sectors and shares and display time-series graph as well as the result of egcm to check the co-integration for the selected pair.